+++ cmd/system-probe/subcommands/runtime/security_profile.go
@@ -21,9 +21,9 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/secrets"
 	secagent "github.com/DataDog/datadog-agent/pkg/security/agent"
 	"github.com/DataDog/datadog-agent/pkg/security/proto/api"
-	timeResolver "github.com/DataDog/datadog-agent/pkg/security/resolvers/time"
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile/profile"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	timeresolver "github.com/DataDog/datadog-agent/pkg/util/time"
 )
 
 type securityProfileCliParams struct {
@@ -161,7 +161,7 @@ func printActivityTreeStats(prefix string, msg *api.ActivityTreeStatsMessage) {
 }
 
 func printSecurityProfileMessage(msg *api.SecurityProfileMessage) {
-	timeResolver, err := timeResolver.NewResolver()
+	timeResolver, err := timeresolver.NewResolver()
 	if err != nil {
 		fmt.Printf("can't get new time resolver: %v\n", err)
 		return
